fix: move util to correct file
